+++ x/bank/internal/types/key.go
@@ -1,8 +1,20 @@
 package types
 
 const (
-	// module name
-	ModuleName   = "bank"
+	// ModuleName defines the module name
+	ModuleName = "bank"
+
+	// StoreKey defines the primary module store key
+	StoreKey = ModuleName
+
+	// RouterKey defines the module's message routing key
+	RouterKey = ModuleName
+
+	// QuerierRoute defines the module's query routing key
 	QuerierRoute = ModuleName
-	StoreKey     = ModuleName
+)
+
+// KVStore key prefixes
+var (
+	BalancesPrefix = []byte("balances")
 )